// src/models/FullProjectModel.ts
export default interface FullProjectModel {
  projectId: number;
  rsdp: string;
  name: string;
  number: string;
  description?: any;
  startDate: any;
  duration?: any;
  dueDate?: any;
  roadLength?: number;
  createdBy?: any;
  lastEditedBy?: string;
  creatingDirectorate?: any;
  createdOn?: any;
  lastModifiedOn?: any;
}